1.拆分transform相关定义到position+scale,移除rotation
2.增强sprite组件,现在分为*纹理*组件和*渲染*组件

3.添加精灵特效(位置/缩放/颜色)

4.添加更多注释说明

+++ src-new/component/sprite_animation.hpp
@@ -13,9 +13,10 @@
 
 namespace pd::component::sprite_animation
 {
-	// sprite
+	// 精灵动画 == 精灵 + 每帧切换
+	// sprite --> 精灵动画精灵
 
-	// 所有帧信息
+	// 精灵动画所有帧信息
 	class Frames final
 	{
 	public:
@@ -24,8 +25,8 @@ namespace pd::component::sprite_animation
 		public:
 			// sprite::Texture
 			manager::texture_handler texture;
-			// sprite::Position
-			sf::Vector2i position;
+			// sprite::TexturePosition
+			sf::Vector2f position;
 		};
 
 		std::vector<Frame> frames;
